pyproject.toml
[tool.poetry]
name = "glucosemonitorapi"
version = "0.1.0"
description = "GlucoseMonitorAPI is a Django-based backend service for storing and retrieving blood glucose levels. It provides endpoints for querying, filtering, and exporting glucose data. Features include pagination, sorting, data pre-population via API, and export to JSON, CSV, and Excel."
authors = ["Ruslan Ibragimov <ruslan.ibragimov.252@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "3.15.2"
django = "4.2.13"
psycopg2 = "2.9.9"
python-decouple = "^3.8"
pandas = "^2.2.2"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.8.0"
pytest = "^8.2.2"
faker = "^26.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
